fix a bug
Invalid update: invalid number of rows in section 0. The number of rows
contained in an existing section after the update (2) must be equal to
the number of rows contained in that section before the update (5),
plus or minus the number of rows inserted or deleted from that section
(0 inserted, 1 deleted) and plus or minus the number of rows moved into
or out of that section (0 moved in, 0 moved out).
FlashCardCreator-[MasterViewController deleteCurrentCard:]