fix a bug
From client:
While creating a card the screen text and picture area moved up the
screen within the question and answer cards. A picture of the pages is
attached.
The App had to be shut down to enable the screen to return to normal
and allow me proceed in creating the cards.