Text to speech changes.
place about a 2 sec delay in the text to speech starting on the Q & A
of each cards created as when you access the card it starts straight
away and I think there should be a slight pause.
I the text to speech still seem to be reading quite fast, slow it down
a bit more.