fix a vertical alignment problem, see background
hen entering text or symbol's and having select the alignment you
require. When playing the card all is well, but when you return to
editing page the text has moved up to the top of the text box as per
the below images.The is happening on new and old card packs.