video volume too low, see background
background:  When you play the small video on the card the sound is
good, but once you select the large screen the sound is reduced and on
moving back to the small video screen and play the video the sound is
also reduced. I’m only able to get the sound back by moving back to the
edit card where the sound is still reduced and then by play the card
again, the text to speech and video sound return. I play the video it
all good, but not after I play the large screen very poor sound.